+++ classes/Login.php
@@ -230,7 +230,7 @@ class Login
 
         $username = $this->validateField('username', $data['username']);
 
-        $file = CompiledYamlFile::instance($this->grav['locator']->findResource('account://' . $username . YAML_EXT,
+        $file = CompiledYamlFile::instance($this->grav['locator']->findResource('account://' . mb_strtolower($username) . YAML_EXT,
             true, true));
 
         // Create user object and save it
@@ -423,9 +423,6 @@ class Login
             $author
         ]);
         $to = $user->email;
-
-
-
         $sent = EmailUtils::sendEmail($subject, $content, $to);
 
         if ($sent < 1) {
@@ -452,12 +449,14 @@ class Login
         if (!$this->rememberMe) {
             /** @var Config $config */
             $config = $this->grav['config'];
+            $cookieName = $config->get('plugins.login.rememberme.name');
+            $timeout = $config->get('plugins.login.rememberme.timeout');
 
             // Setup storage for RememberMe cookies
-            $storage = new TokenStorage;
+            $storage = new TokenStorage('user://data/rememberme', $timeout);
             $this->rememberMe = new RememberMe($storage);
-            $this->rememberMe->setCookieName($config->get('plugins.login.rememberme.name'));
-            $this->rememberMe->setExpireTime($config->get('plugins.login.rememberme.timeout'));
+            $this->rememberMe->setCookieName($cookieName);
+            $this->rememberMe->setExpireTime($timeout);
 
             // Hardening cookies with user-agent and random salt or
             // fallback to use system based cache key